dev/requirements.txt
fastapi[standard]==0.115.4
passlib==1.7.4
uvicorn==0.32.0
python-dotenv==1.0.1
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
PyMySQL==1.1.1
mysql-connector-python==9.1.0
PyJWT==2.9.0
FastAPI-SQLAlchemy==0.2.1
starlette==0.41.2
requests==2.32.3
redis==5.2.0
cryptography==44.0.0
httpx
mysqlclient